refactor(menus): split InspectorPanel.tsx into smaller files
`InspectorPanel.tsx` had grown to 2030 lines mixing the dispatcher, the
menu-level inspector, the button inspector, three scene-node inspectors,
shared style-section widgets, and pure helper/diagnostic logic in one file
(#48). Before splitting, extended `SceneEditor.test.tsx`'s existing
InspectorPanel tests with characterization coverage for the previously
untested areas: text/image/shape node panels, the generic-node fallback,
the menu-level background-tab switcher, ButtonStyleSection's state-tab
switching, action serialise/parse round-tripping, and the diagnostics list
— run green against the pre-split code first.

Split into:
- `InspectorCollapsibleSection.tsx` — the generic chevron-toggle wrapper
- `InspectorSharedFields.tsx` — `ActionOptions` and `HighlightColourFields`
- `ButtonStyleSection.tsx` / `TextStyleSection.tsx` — per-state button style
  and typography widgets
- `SceneNodeInspectors.tsx` — text/image/shape/generic node panels
- `MenuLevelInspector.tsx` — the no-selection menu-level panel (diagnostics,
  background, display aspect, button audit, compile policy, CLUT palette,
  auto-nav, render preview)
- `ButtonInspector.tsx` — the button-selected panel
- `inspectorHelpers.ts` — title/subtitle copy and action <-> string codecs
- `inspectorDiagnostics.ts` — the DVD-safety diagnostic checks

Also deleted `InspectorPanel.tsx`'s local `DEFAULT_BUTTON_STYLE_MAP`/
`DEFAULT_TEXT_STYLE` duplicates and switched `ButtonInspector.tsx` to import
them from the existing `menuDefaults.ts` (already the source of truth since
the prior `MenusPage.tsx` split) — a pure dedup, same values.

`InspectorPanel.tsx` itself is now ~240 lines: just `InspectorPanelProps`
and the selection-dispatching `InspectorPanel` component.

Verified with `tsc --noEmit`, `pnpm test:js` (94 passing — same count
pre- and post-split), `pnpm format:check`, and `pnpm build`.
